fix: move shared components base-color, color-edit, color-channel-slider to express/code/scripts/color-shared/components
