+++ express/code/scripts/color-shared/components/base-color/index.js
@@ -1,4 +1,4 @@
-import { LitElement, html, nothing } from '../../../deps/lit-all.min.js';
+import { LitElement, html, nothing } from '../../../../libs/deps/lit-all.min.js';
 import { style } from './styles.css.js';
 import {
   hexToRGB,
@@ -8,9 +8,9 @@ import {
   hsbToHSL,
   rgbToLab,
   labToRGB,
-} from '../../utils/ColorConversions.js';
-import { loadMenu, loadButton, loadColorArea, loadColorSlider, loadTextfield } from '../../../../scripts/color-shared/spectrum/load-spectrum.js';
-import { loadColorTokens } from '../../../../scripts/color-shared/utils/loadColorTokens.js';
+} from '../../../../libs/color-components/utils/ColorConversions.js';
+import { loadMenu, loadButton, loadColorArea, loadColorSlider, loadTextfield } from '../../spectrum/load-spectrum.js';
+import { loadColorTokens } from '../../utils/loadColorTokens.js';
 import '../color-channel-slider/index.js';
 
 const COLOR_MODES = ['HEX', 'RGB', 'HSB', 'Lab'];
