+++ express/code/scripts/color-shared/spectrum/index.js
@@ -40,7 +40,6 @@ export { createExpressSwatchGroup } from './components/express-swatch-group.js';
 export { createExpressColorArea } from './components/express-color-area.js';
 export { createExpressColorSlider } from './components/express-color-slider.js';
 export { createExpressSlider } from './components/express-slider.js';
-export { default as ColorChannelSlider } from '../../../libs/color-components/components/color-channel-slider/index.js';
 
 // ── Utilities ────────────────────────────────────────────────────────
 export { createThemeWrapper, wrapInTheme } from './utils/theme.js';
